arrumando os dtos respostas

+++ src/main/java/com/example/ControleAlmoxarifado/dto/fornecedor/CriacaoFornecedorRespostaDTO.java
@@ -1,7 +1,7 @@
 package com.example.ControleAlmoxarifado.dto.fornecedor;
 
 public record CriacaoFornecedorRespostaDTO(
-        int id,
+        Long id,
         String nome,
         String cnpj
 ) {
